Separating enums to files and removing unused
